@@ -21,8 +21,8 @@
     "dependencies": {
         "@fontsource/inter": "^4.5.15",
         "@types/mime-types": "^2.1.1",
-        "@types/node": "^18.15.10",
-        "@types/react": "^18.0.30",
+        "@types/node": "^18.15.11",
+        "@types/react": "^18.0.31",
         "@types/react-dom": "^18.0.11",
         "autoprefixer": "^10.4.14",
         "cross-env": "^7.0.3",
@@ -34,21 +34,21 @@
         "postcss": "^8.4.21",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
-        "tailwindcss": "^3.2.7",
+        "tailwindcss": "^3.3.0",
         "typescript": "^5.0.2"
     },
     "devDependencies": {
-        "@commitlint/cli": "^17.5.0",
+        "@commitlint/cli": "^17.5.1",
         "@commitlint/config-conventional": "^17.4.4",
         "@typescript-eslint/eslint-plugin": "^5.57.0",
         "@typescript-eslint/parser": "^5.57.0",
-        "eslint": "^8.36.0",
+        "eslint": "^8.37.0",
         "eslint-config-next": "^13.2.4",
         "eslint-plugin-prettier": "^4.2.1",
         "eslint-plugin-react": "^7.32.2",
         "eslint-plugin-security": "^1.7.1",
         "husky": "^8.0.3",
         "prettier": "^2.8.7",
-        "prettier-plugin-tailwindcss": "^0.2.5"
+        "prettier-plugin-tailwindcss": "^0.2.6"
     }
 }
