@@ -15,7 +15,7 @@ export default function ProjectImages(props: Prop): JSX.Element {
                         src={`/images/projects/${project.slug}/image${image + 1}.png`}
                         alt={`${project.name} image`}
                         loading="lazy"
-                        className="mt-0 mr-4 mb-5 ml-0 max-h-[30rem] max-w-full rounded shadow-xl"
+                        className="mb-5 ml-0 mr-4 mt-0 max-h-[30rem] max-w-full rounded shadow-xl"
                     />
                 </li>
             ))}
